webgl

Module webgl_thread

Structs

FramebufferRebindingFlags 🔒

FramebufferRebindingInfo 🔒

GLContextData 🔒

GLState

WebGLContextInfo 🔒

Helper struct to store cached WebGLContext information.

WebGLImpl

WebGL Commands Implementation

WebGLThread 🔒

A WebGLThread manages the life cycle and message multiplexing of a set of WebGLContexts living in the same thread.

WebGLThreadInit 🔒

The data required to initialize an instance of the WebGLThread type.

Constants

ANGLE_NAME_PREFIX 🔒

ANGLE adds a _u prefix to variable names:

SAFE_VIEWPORT_DIMS 🔒

Traits

SurfmanContextAttributeFlagsConvert 🔒

ToSurfmanVersion 🔒

Functions

clamp_viewport 🔒

flip_pixels_y 🔒

Flips the pixels in the Vec on the Y axis.

from_name_in_compiled_shader 🔒

Removes _u prefix from variable names

image_to_tex_image_data 🔒

Translates an image in rgba8 (red in the first byte) format to the format that was requested of TexImage.

map_dot_separated 🔒

native_uniform_location 🔒

premultiply_inplace 🔒

prepare_pixels 🔒

to_name_in_compiled_shader 🔒

Adds _u prefix to variable names

Type Aliases

GLint 🔒

WebGLContextBusyMap

A map which tracks whether a given WebGL context is “busy” ie whether WebRender has currently taken a surface from its [SwapChain] for rendering purposes. Contexts will only be deleted once no WebRender instance is using it for rendering. This ensures that all Surfman Surfaces can be released properly on the WebGLThread.