webgl::webgl_thread

Type Alias WebGLContextBusyMap

pub type WebGLContextBusyMap = Arc<RwLock<HashMap<WebGLContextId, usize>>>;

A map which tracks whether a given WebGL context is “busy” ie whether WebRender has currently taken a surface from its [SwapChain] for rendering purposes. Contexts will only be deleted once no WebRender instance is using it for rendering. This ensures that all Surfman Surfaces can be released properly on the WebGLThread.

Aliased Type

pub struct WebGLContextBusyMap { /* private fields */ }